pyproject.toml
[tool.poetry]
name = "hexlet-code"
version = "0.1.0"
description = "generating differences between two files"
classifiers = [
    "Topic :: Software Development :: Build Tools",
    "Topic :: Software Development :: Libraries :: Python Modules",
    "Programming Language :: Python :: 3",
]
authors = ["Nella <nella611@yandex.ru>"]
packages = [
  { include = "gendiff" },
]

[[tool.poetry.source]]
name = 'hexlet-code'
url = 'https://github.com/Nella611/python-project-lvl2'

[tool.poetry.dependencies]
python = "^3.8"
PyYAML = "^6.0"

[tool.poetry.dev-dependencies]
flake8 = "^4.0.1"
pytest = "^6.2.5"
pytest-cov = "^3.0.0"
coverage = "^6.2"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"

[tool.poetry.scripts]
gendiff = "gendiff.scripts.gendiff:main"
